\name{DataForPREDA2dataframe}
\alias{DataForPREDA2dataframe}
\alias{DataForPREDA2dataframe,DataForPREDA-method}
\title{
extract data and annotations as a dataframe
}
\description{
extract data and annotations as a dataframe with probeids as rownames
}
\usage{
DataForPREDA2dataframe(.Object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.Object}{
 An object of class DataForPREDA
}
}
\details{
extract data and annotations as a dataframe with probeids as rownames
}
\value{
a dataframe with probeids as rownames
}

