\name{DataForPREDA2GenomicAnnotationsForPREDA}
\alias{DataForPREDA2GenomicAnnotationsForPREDA}
\alias{DataForPREDA2GenomicAnnotationsForPREDA,DataForPREDA-method}
\title{
extract a GenomicAnnotationsForPREDA object from a data DataForPREDA object
}
\description{
extract a GenomicAnnotationsForPREDA object from a data DataForPREDA object
}
\usage{
DataForPREDA2GenomicAnnotationsForPREDA(.Object)
}
\arguments{
  \item{.Object}{
an object of class DataForPREDA
}
}
\details{
extract a GenomicAnnotationsForPREDA object from a data DataForPREDA object
}
\value{
a GenomicAnnotationsForPREDA object
}
