\name{DataForPREDA2StatisticsForPREDA}
\alias{DataForPREDA2StatisticsForPREDA}
\alias{DataForPREDA2StatisticsForPREDA,DataForPREDA-method}
\title{
extract a StatisticsForPREDA object from a data DataForPREDA object
}
\description{
extract a StatisticsForPREDA object from a data DataForPREDA object
}
\usage{
DataForPREDA2StatisticsForPREDA(.Object)
}
\arguments{
  \item{.Object}{
 a data DataForPREDA object
}
}
\details{
extract a StatisticsForPREDA object from a data DataForPREDA object
}
\value{
a StatisticsForPREDA object
}

