\name{GenomicAnnotations2dataframe}
\alias{GenomicAnnotations2dataframe}
\alias{GenomicAnnotations2dataframe,GenomicAnnotations-method}
\alias{GenomicAnnotations2dataframe,GenomicAnnotationsForPREDA-method}
\title{
extracts annotations as a dataframe
}
\description{
extracts annotations as a dataframe with probeids as rownames
}
\usage{
GenomicAnnotations2dataframe(.Object)
}
\arguments{
  \item{.Object}{
A GenomicAnnotations object
}
}
\details{
extract annotations as a dataframe with probeids as rownames
}
\value{
a dataframe with probeids as rownames
}

