\name{GenomicAnnotationsForPREDA2dataframe}
\alias{GenomicAnnotationsForPREDA2dataframe}
\alias{GenomicAnnotationsForPREDA2dataframe,GenomicAnnotationsForPREDA-method}
\title{
extract annotations as a dataframe
}
\description{
extract annotations as a dataframe with probeids as rownames
}
\usage{
GenomicAnnotationsForPREDA2dataframe(.Object)
}
\arguments{
  \item{.Object}{
an object of class GenomicAnnotationsForPREDA
}
}
\details{
extract annotations from an object of class GenomicAnnotationsForPREDA
as a dataframe with probeids as rownames

}
\value{
a dataframe with probeids as rownames
}
