\name{GenomicAnnotationsForPREDA2GenomicAnnotations}
\alias{GenomicAnnotationsForPREDA2GenomicAnnotations}
\alias{GenomicAnnotationsForPREDA2GenomicAnnotations,GenomicAnnotationsForPREDA-method}
\title{
extract the GenomicAnnotations object from the GenomicAnnotationsForPREDA object
}
\description{
extract the GenomicAnnotations object from the GenomicAnnotationsForPREDA object
}
\usage{
GenomicAnnotationsForPREDA2GenomicAnnotations(.Object)
}
\arguments{
  \item{.Object}{
an object of class GenomicAnnotationsForPREDA
}
}
