\name{GenomicAnnotationsForPREDA2PREDAResults}
\alias{GenomicAnnotationsForPREDA2PREDAResults}
\alias{GenomicAnnotationsForPREDA2PREDAResults,GenomicAnnotationsForPREDA-method}
\title{
add PREDA results information to genomic annotatations creating a PREDAResults object
}
\description{
add PREDA results information to genomic annotatations creating a PREDAResults object
}
\usage{
# GenomicAnnotationsForPREDA2PREDAResults(.Object, analysesNames, testedTail, smoothStatistic, pvalue, qvalue)

GenomicAnnotationsForPREDA2PREDAResults(.Object, ...)
}
\arguments{
  \item{.Object}{
An object of class GenomicAnnotationsForPREDA
}
  \item{\dots}{
   See below
   \describe{
    \item{analysesNames:}{
    analysesNames as in PREDAResults object
    }
      \item{testedTail:}{
    testedTail as in PREDAResults object
    }
      \item{smoothStatistic:}{
    smoothStatistic as in PREDAResults object
    }
      \item{pvalue:}{
    pvalue as in PREDAResults object
    }
      \item{qvalue:}{
    qvalue as in PREDAResults object
    }
  }
}
}
