\name{GenomicRegionsChrNumber}
\alias{GenomicRegionsChrNumber}
\alias{GenomicRegionsChrNumber,GenomicRegions-method}
\title{
determine the number of chromosomes with genomic regions
}
\description{
determine the number of chromosomes with genomic regions
}
\usage{
GenomicRegionsChrNumber(.Object)
}
\arguments{
  \item{.Object}{
An object of class GenomicRegions
}
}
