\name{GenomicRegionsNumber}
\alias{GenomicRegionsNumber}
\alias{GenomicRegionsNumber,GenomicRegions-method}
\title{
determine the number of genomic regions
}
\description{
determine the number of genomic regions
}
\usage{
GenomicRegionsNumber(.Object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.Object}{
An object of class GenomicRegions
}
}
