\name{getStatisticByName}
\alias{getStatisticByName}
\alias{getStatisticByName,StatisticsForPREDA-method}
\title{
extract data for individual analyses using the analysis name
}
\description{
extract data for individual analyses using the analysis name
}
\usage{
# getStatisticByName(.Object, analysisName)

getStatisticByName(.Object, ...)
}
\arguments{
  \item{.Object}{
An object of class StatisticsForPREDA
}
  \item{\dots}{
  See below
     \describe{
        \item{analysisName:}{
        Character name of the analysis to be returned
        }
      }
}
}
