\name{PREDADataAndResults2dataframe}
\alias{PREDADataAndResults2dataframe}
\alias{PREDADataAndResults2dataframe,PREDADataAndResults-method}
\title{
extract data and annotations as a dataframe with probeids as rownames
}
\description{
extract data and annotations as a dataframe with probeids as rownames
}
\usage{
PREDADataAndResults2dataframe(.Object)
}
\arguments{
  \item{.Object}{
An object of class PREDADataAndResults
}
}
