\name{PREDAResults2dataframe}
\alias{PREDAResults2dataframe}
\alias{PREDAResults2dataframe,PREDAResults-method}
\title{
extact preda results statistics as a data frame object
}
\description{
extact preda results statistics as a data frame object
}
\usage{
PREDAResults2dataframe(.Object)
}
\arguments{
  \item{.Object}{
An object of class PREDAResults
}
}
