\name{PREDAResults2PREDADataAndResults}
\alias{PREDAResults2PREDADataAndResults}
\alias{PREDAResults2PREDADataAndResults,PREDAResults-method}
\title{
merge PREDAResults and input statistics to create a PREDADataAndResults object
}
\description{
merge PREDAResults and input statistics to create a PREDADataAndResults object
}
\usage{
# PREDAResults2PREDADataAndResults(.Object, statistic)

PREDAResults2PREDADataAndResults(.Object, ...)
}
\arguments{
  \item{.Object}{
An object of class PREDAResults
}
  \item{\dots}{
   See below
   \describe{
   \item{statistic:}{
A matrix containing input statistics
}
  }
}
}
