\name{StatisticsForPREDA2dataframe}
\alias{StatisticsForPREDA2dataframe}
\alias{StatisticsForPREDA2dataframe,StatisticsForPREDA-method}
\title{
extract data as a dataframe with probeids as rownames
}
\description{
extract data as a dataframe with probeids as rownames
}
\usage{
StatisticsForPREDA2dataframe(.Object)
}
\arguments{
  \item{.Object}{
An object of class StatisticsForPREDA
}
}
